src/cart_wheel/channel.py
"""Channel management - repodata generation for conda channels."""

from __future__ import annotations

import asyncio
from pathlib import Path

from rattler.index import index_fs


def index_channel(channel_dir: Path) -> None:
    """Generate repodata.json for each subdir in the channel.

    Uses py-rattler's index_fs to generate proper repodata.json files
    with support for sharded repodata and compression.

    Args:
        channel_dir: Root directory of the conda channel
    """
    asyncio.run(_index_channel_async(channel_dir))


async def _index_channel_async(channel_dir: Path) -> None:
    """Async implementation of channel indexing."""
    await index_fs(
        channel_directory=channel_dir,
        target_platform=None,  # Index all platforms
        write_zst=True,  # Generate compressed repodata
        write_shards=False,  # Skip sharded repodata for simplicity
        force=False,  # Only re-index if needed
    )


def prune_channel(
    channel_dir: Path,
    state_dir: Path,
    keep_versions: int,
) -> list[Path]:
    """Remove old versions from channel.

    Args:
        channel_dir: Root directory of the conda channel
        state_dir: Directory containing state files
        keep_versions: Number of versions to keep per package

    Returns:
        List of removed file paths
    """
    # This is a placeholder - full implementation would need to:
    # 1. Group packages by name
    # 2. Sort by version
    # 3. Remove all but keep_versions newest
    # 4. Update state files
    # For now, return empty list
    return []
