pyproject.toml
[project]
name = "cart-wheel"
version = "0.1.0"
description = "Convert Python wheels to conda packages"
requires-python = ">=3.10"
dependencies = [
    "zstandard>=0.21.0",
    "pkginfo>=1.12.1",
    "packaging>=25.0",
    "stream-unzip>=0.0.95,<0.0.100",
    "py-rattler>=0.20.0",
    "httpx>=0.28.1",
    "hishel[httpx]>=1.1.7",
    "rich>=13.0.0",
    "InquirerPy>=0.3.4"
]

[project.scripts]
cart-wheel = "cart_wheel.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cart_wheel"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]
preview = ["pixi-build"]

[tool.pixi.package.build]
backend = { name = "pixi-build-python", version = "0.4.*" }

[tool.pixi.package.host-dependencies]
hatchling = "*"

[tool.pixi.package.run-dependencies]
zstandard=">0.21.0"
pkginfo = ">=1.12.1"
packaging = ">=25.0"
stream-unzip = ">=0.0.95,<0.0.100"
py-rattler = "0.20.*"
httpx = ">=0.28.1"
rich = ">=13.0.0"

[tool.pixi.dependencies]
cart-wheel = { path = "." }

[tool.pixi.tasks]
convert = "cart-wheel convert"
sync = "cart-wheel sync"
status = "cart-wheel status"
add-package = "cart-wheel add"

[tool.pixi.feature.test.dependencies]
pytest = ">=8.0"

[tool.pixi.feature.test.tasks]
test = "pytest tests -v"

[tool.pixi.environments]
default = { features = [], solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[tool.pytest.ini_options]
testpaths = ["tests"]
